requirements.txt
# KRATOS - Kubernetes Runtime Agentic Operating System
# Core dependencies
streamlit>=1.31.0
pyautogen>=0.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
asyncio-mqtt>=0.13.0
requests>=2.31.0

# Kubernetes and Azure dependencies
kubernetes>=28.1.0
azure-cli-core>=2.55.0
azure-identity>=1.15.0
azure-mgmt-containerservice>=26.0.0
azure-mgmt-resource>=23.1.0

# MCP and communication
websockets>=12.0
httpx>=0.26.0
aiohttp>=3.9.0

# Logging and monitoring
structlog>=23.2.0
rich>=13.7.0
watchdog>=3.0.0

# Data handling
pandas>=2.1.0
numpy>=1.26.0
plotly>=5.17.0

# Development and testing
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.8.0

# Optional: Docker support
docker>=7.0.0